+++ test/test_helper.exs
@@ -34,6 +34,8 @@ ExUnit.start(
     # :issuing_holder_log,
     # :issuing_invoice,
     # :issuing_invoice_log,
+    # :issuing_billing_invoice,
+    # :issuing_billing_transaction,
     # :issuing_purchase,
     # :issuing_purchase_log,
     # :issuing_transaction,